requirements.txt
django == 3.1.1
djangorestframework == 3.11.1
pillow == 7.2.0
psycopg2-binary == 2.8.6
django-cors-headers==3.5.0
django-filter==2.4.0
django-import-export
djangorestframework-jwt 

